+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `fast_init` option
 - Improved performance of loading models
 - Fixed error with `aggregation_strategy` option